Added smart constructor selection to void Throwable constructors
When instantiating the exception with a Throwable cause, the exception might get logged in the application with the cause used in ReflectionErrorDecoder. This created a weird looking log with a cause that would only be a Throwable with no error message and would be misleading.

This commit add a smart ordering of the constructors to favor constructors with no Throwable argument. It also adds a clear error message for the user indicating that this Throwable was used only for instantiation.